+++ hpOneView/image_streamer/resources/artifact_bundles.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright [2019] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2020] Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -24,92 +24,21 @@ from future import standard_library
 standard_library.install_aliases()
 
 
-from hpOneView.resources.resource import ResourceClient, extract_id_from_uri
+from hpOneView.resources.resource import Resource, ResourceFileHandlerMixin, extract_id_from_uri
 
 
-class ArtifactBundles(object):
+class ArtifactBundles(ResourceFileHandlerMixin, Resource):
 
     URI = '/rest/artifact-bundles'
-    DEPLOYMENT_GROUPS_URI = '/rest/deployment-groups/'
     BACKUPS_PATH = '/rest/artifact-bundles/backups'
     BACKUP_ARCHIVE_PATH = '/rest/artifact-bundles/backups/archive'
-    STOP_CREATION_PATH = '/stopArtifactCreate'
     DOWNLOAD_PATH = '/rest/artifact-bundles/download'
 
-    DEFAULT_VALUES = {
-        '300': {"type": "ArtifactsBundle"}
-    }
-
-    def __init__(self, con):
-        self._client = ResourceClient(con, self.URI)
+    def __init__(self, connection, data=None):
+        super(ArtifactBundles, self).__init__(connection, data)
         self.__default_values = {
-            'type': 'ArtifactsBundle',
+            'type': 'ArtifactsBundle'
         }
-
-    def get_all(self, start=0, count=-1, filter='', sort=''):
-        """
-        Gets a list of Artifacts Bundle based on optional sorting and filtering, and constrained by start and count
-        parameters.
-
-        Args:
-            start:
-                The first item to return, using 0-based indexing.
-                If not specified, the default is 0 - start with the first available item.
-            count:
-                The number of resources to return. A count of -1 requests all items.
-
-                The actual number of items in the response might differ from the requested
-                count if the sum of start and count exceeds the total number of items.
-            filter (list or str):
-                A general filter/query string to narrow the list of items returned. The
-                default is no filter; all resources are returned.
-            sort:
-                The sort order of the returned data set. By default, the sort order is based
-                on create time with the oldest entry first.
-
-        Returns:
-            list: A list of Artifacts Bundle.
-        """
-        return self._client.get_all(start, count, filter=filter, sort=sort)
-
-    def get(self, id_or_uri):
-        """
-        Retrieves the overview details for the selected Artifact Bundle as per the selected attributes.
-
-        Args:
-            id_or_uri: ID or URI of the Artifact Bundle.
-
-        Returns:
-            dict: The Artifact Bundle.
-        """
-        return self._client.get(id_or_uri)
-
-    def get_by(self, field, value):
-        """
-        Gets all of the Artifacts Bundle resources that match the filter.
-
-        The search is case-insensitive.
-
-        Args:
-            field: Field name to filter.
-            value: Value to filter.
-
-        Returns:
-            list: The Artifacts Bundle.
-        """
-        return self._client.get_by(field, value)
-
-    def get_by_name(self, name):
-        """
-        Gets an Artifact Bundle by name.
-
-        Args:
-            name: Name of the Artifact Bundle.
-
-        Returns:
-            dict: The Artifact Bundle.
-        """
-        return self._client.get_by_name(name)
 
     def get_all_backups(self):
         """
@@ -118,49 +47,54 @@ class ArtifactBundles(object):
         Returns:
             list: A list of Backups for Artifacts Bundle.
         """
-        return self._client.get_collection(id_or_uri=self.BACKUPS_PATH)
+        return self._helper.get_collection(uri=self.BACKUPS_PATH)
 
-    def get_backup(self, id_or_uri):
+    def get_backup(self, uri):
         """
         Get the details for the backup from an Artifact Bundle.
 
         Args:
-            id_or_uri: ID or URI of the Artifact Bundle.
+            uri: URI of the Artifact Bundle.
 
         Returns:
             Dict: Backup for an Artifacts Bundle.
         """
-        uri = self.BACKUPS_PATH + '/' + extract_id_from_uri(id_or_uri)
-        return self._client.get(id_or_uri=uri)
+        uri = self.BACKUPS_PATH + '/' + extract_id_from_uri(uri)
+        return super(ArtifactBundles, self).get_by_uri(uri)
 
-    def download_archive_artifact_bundle(self, id_or_uri, file_path):
+    def download_archive(self, path, uri=None):
         """
-        Downloads an archive for the Artifact Bundle.
+        Downloads backup archive for the Artifact Bundle.
 
         Args:
-            id_or_uri: ID or URI of the Artifact Bundle.
-            file_path(str): Destination file path.
+            path(str): Destination file path.
+            uri: URI of the Artifact Bundle.
 
         Returns:
             bool: Successfully downloaded.
         """
+        if not uri:
+            uri = self.data['uri']
 
-        uri = self.BACKUP_ARCHIVE_PATH + '/' + extract_id_from_uri(id_or_uri)
-        return self._client.download(uri, file_path)
+        download_uri = self.BACKUP_ARCHIVE_PATH + '/' + extract_id_from_uri(uri)
+        return super(ArtifactBundles, self).download(download_uri, path)
 
-    def download_artifact_bundle(self, id_or_uri, file_path):
+    def download(self, path, uri=None):
         """
         Download the Artifact Bundle.
 
         Args:
-            id_or_uri: ID or URI of the Artifact Bundle.
-            file_path(str): Destination file path.
+            uri: URI of the Artifact Bundle.
+            path(str): Destination file path.
 
         Returns:
             bool: Successfully downloaded.
         """
-        uri = self.DOWNLOAD_PATH + '/' + extract_id_from_uri(id_or_uri)
-        return self._client.download(uri, file_path)
+        if not uri:
+            uri = self.data['uri']
+
+        download_uri = self.DOWNLOAD_PATH + '/' + extract_id_from_uri(uri)
+        return super(ArtifactBundles, self).download(download_uri, path)
 
     def create_backup(self, resource, timeout=-1):
         """
@@ -176,7 +110,7 @@ class ArtifactBundles(object):
         Returns:
             dict: A Deployment Group associated with the Artifact Bundle backup.
         """
-        return self._client.create(resource, uri=self.BACKUPS_PATH, timeout=timeout)
+        return super(ArtifactBundles, self).create(resource, uri=self.BACKUPS_PATH, timeout=timeout)
 
     def upload_bundle_from_file(self, file_path):
         """
@@ -188,80 +122,29 @@ class ArtifactBundles(object):
         Returns:
             dict: Artifact bundle.
         """
-        return self._client.upload(file_path)
+        return super(ArtifactBundles, self).upload(file_path)
 
-    def upload_backup_bundle_from_file(self, file_path, deployment_groups_id_or_uri):
+    def upload_backup_bundle_from_file(self, file_path, deployment_group_uri):
         """
         Restore an Artifact Bundle from a backup file.
 
         Args:
             file_path (str): The File Path to restore the Artifact Bundle.
-            deployment_groups_id_or_uri: ID or URI of the Deployment Groups.
+            deployment_group_uri: URI of the Deployment Groups.
 
         Returns:
             dict: Deployment group.
         """
-        deployment_groups_uri = deployment_groups_id_or_uri
+        uri = self.BACKUP_ARCHIVE_PATH + "?deploymentGrpUri=" + deployment_group_uri
+        return super(ArtifactBundles, self).upload(file_path, uri)
 
-        if self.DEPLOYMENT_GROUPS_URI not in deployment_groups_id_or_uri:
-            deployment_groups_uri = self.DEPLOYMENT_GROUPS_URI + deployment_groups_id_or_uri
-
-        uri = self.BACKUP_ARCHIVE_PATH + "?deploymentGrpUri=" + deployment_groups_uri
-
-        return self._client.upload(file_path, uri)
-
-    def create(self, resource, timeout=-1):
-        """
-        Creates an Artifact Bundle.
-
-        Args:
-            resource (dict): Object to create.
-            timeout:
-                Timeout in seconds. Waits for task completion by default. The timeout does not abort the operation
-                in OneView, it just stops waiting for its completion.
-
-        Returns:
-            dict: Created resource.
-        """
-        return self._client.create(resource, timeout=timeout)
-
-    def delete(self, resource, timeout=-1):
-        """
-        Deletes an Artifact Bundle.
-
-        Args:
-            resource(str, dict):
-                Accept either the resource id  or the entire resource.
-            timeout:
-                Timeout in seconds. Waits for task completion by default. The timeout does not abort the operation in
-                OneView, it just stops waiting for its completion.
-
-        Returns:
-            bool: Indicates if the resource was successfully deleted.
-        """
-        return self._client.delete(resource, timeout=timeout)
-
-    def update(self, resource, timeout=-1):
-        """
-        Updates only name for the Artifact Bundle.
-
-        Args:
-            resource (dict): Object to update.
-            timeout:
-                Timeout in seconds. Waits for task completion by default. The timeout does not abort the operation
-                in OneView, it just stops waiting for its completion.
-
-        Returns:
-            dict: Updated resource.
-        """
-        return self._client.update(resource, timeout=timeout, default_values=self.DEFAULT_VALUES)
-
-    def extract_bundle(self, resource, timeout=-1):
+    def extract(self, resource=None, uri=None, timeout=-1):
         """
         Extracts the existing bundle on the appliance and creates all the artifacts.
 
         Args:
             resource (dict): Artifact Bundle to extract.
+            uri: URI of artifact bundle
             timeout:
                 Timeout in seconds. Waits for task completion by default. The timeout does not abort the operation in
                 OneView, it just stops waiting for its completion.
@@ -269,14 +152,19 @@ class ArtifactBundles(object):
         Returns:
             dict: The Artifact Bundle.
         """
-        return self._client.update(resource, timeout=timeout, custom_headers={"Content-Type": "text/plain"})
+        if not uri:
+            uri = self.data['uri']
 
-    def extract_backup_bundle(self, resource, timeout=-1):
+        extract_uri = "{}?extract=true&forceImport=true".format(uri)
+        return self._helper.update(resource, extract_uri, timeout=timeout, custom_headers={"Content-Type": "text/plain;charset=UTF-8"})
+
+    def extract_backup(self, resource, uri=None, timeout=-1):
         """
         Extracts the existing backup bundle on the appliance and creates all the artifacts.
 
         Args:
             resource (dict): Deployment Group to extract.
+            uri: URI of artifact backup
             timeout:
                 Timeout in seconds. Waits for task completion by default. The timeout does not abort the operation in
                 OneView, it just stops waiting for its completion.
@@ -284,23 +172,29 @@ class ArtifactBundles(object):
         Returns:
             dict: A Deployment Group associated with the Artifact Bundle backup.
         """
-        return self._client.update(resource, uri=self.BACKUP_ARCHIVE_PATH, timeout=timeout)
+        if not uri:
+            uri = self.data['uri']
 
-    def stop_artifact_creation(self, id_or_uri, task_uri):
+        extract_uri = self.BACKUPS_PATH + '/' + extract_id_from_uri(uri)
+        return self._helper.update(resource, extract_uri, timeout=timeout)
+
+    def stop_artifact_creation(self, task_uri, uri=None):
         """
         Stops creation of the selected Artifact Bundle.
 
         Args:
-            id_or_uri: ID or URI of the Artifact Bundle.
             task_uri: Task URI associated with the Artifact Bundle.
+            uri: URI of the Artifact Bundle.
 
         Returns:
             string:
         """
+        if not uri:
+            uri = self.data['uri']
+
         data = {
             "taskUri": task_uri
         }
 
-        uri = self.URI + '/' + extract_id_from_uri(id_or_uri) + self.STOP_CREATION_PATH
-
-        return self._client.update(data, uri=uri)
+        stop_uri = "{}/stopArtifactCreate".format(uri)
+        return self._helper.update(data, uri=stop_uri)